LPS-111943 Move context and chartState to proper folders
