+++ modules/dxp/apps/analytics/analytics-reports-web/src/main/resources/META-INF/resources/js/AnalyticsReportsApp.js
@@ -14,7 +14,7 @@ import React, {useContext, useState} from 'react';
 
 import Detail from './components/Detail';
 import Main from './components/Main';
-import ConnectionContext from './state/context';
+import ConnectionContext from './context/context';
 import APIService from './utils/APIService';
 import {numberFormat} from './utils/numberFormat';
