+++ modules/dxp/apps/analytics/analytics-reports-web/src/main/resources/META-INF/resources/js/components/TimeSpanSelector.js
@@ -15,7 +15,7 @@ import {ClayTooltipProvider} from '@clayui/tooltip';
 import PropTypes from 'prop-types';
 import React, {useContext} from 'react';
 
-import ConnectionContext from '../state/context';
+import ConnectionContext from '../context/context';
 
 export default function TimeSpanSelector({
 	disabledNextTimeSpan,
