+++ modules/dxp/apps/analytics/analytics-reports-web/src/main/resources/META-INF/resources/js/state/chartState.js
@@ -11,7 +11,7 @@
 
 import {useContext, useReducer} from 'react';
 
-import ConnectionContext from '../state/context';
+import ConnectionContext from '../context/context';
 
 const ADD_DATA_SET_ITEM = 'add-data-key';
 const NEXT_TIME_SPAN = 'next-time-span';
